#101 Calculate rank in ||, post results each 600 blocks. Small fixes
