+++ app/genesis/genesis.go
@@ -9,7 +9,7 @@ import (
 	distr "github.com/cosmos/cosmos-sdk/x/distribution"
 	"github.com/cosmos/cosmos-sdk/x/slashing"
 	"github.com/cosmos/cosmos-sdk/x/stake"
-	"github.com/cybercongress/cyberd/app/types/coin"
+	"github.com/cybercongress/cyberd/types/coin"
 	"github.com/pkg/errors"
 	tmtypes "github.com/tendermint/tendermint/types"
