--- app/types/acc.go
@@ -1,7 +1,0 @@
-package types
-
-import "github.com/cosmos/cosmos-sdk/x/auth"
-
-func NewCyberdAccount() auth.Account {
-	return &auth.BaseAccount{}
-}
